requirements.txt
telethon>=1.24.0
pytz>=2021.1
google-auth>=2.15.0
google-api-python-client>=2.70.0